Client-side website development
Completed the majority of the design for the homepage, search page, event detail page, and navigation. The homepage presents important information about the charity organization. It dynamically displays the list of charity activities using an API, and you can also click the button to view the detailed information of the event. The search page allows for filtering by specific events and can interact effectively with the API. The activity details page displays the detailed information of a specific event, including a registration button. The navigation can smoothly switch to other pages.